fix: Fix wrong types for some projects

+++ src/main/kotlin/com/iyxan23/asperge/sketchware/models/Project.kt
@@ -49,5 +49,5 @@ data class Project(
     // colors =============================
 
     @SerialName("sketchware_ver")
-    val sketchwareVer: Int,
+    val sketchwareVer: Float,
 )
